SAK-50987 webcomponents Refactor rubrics tests into one per component
https://sakaiproject.atlassian.net/browse/SAK-50987
